+++ src/snowmew-core/camera.rs
@@ -53,10 +53,10 @@ fn view_matrix(t: &Matrix4<f32>) -> Matrix4<f32> {
 
 impl Camera {
     /// Create a new camera from a position matrix
-    pub fn new(width: f32, height: f32, transform: Matrix4<f32>) -> Camera {
+    pub fn new(width: uint, height: uint, transform: Matrix4<f32>) -> Camera {
         Camera {
-            width: width,
-            height: height,
+            width: width as f32,
+            height: height as f32,
             transform: transform
         }
     }
